refactor: extract computeDropCoords utility, remove getZoomLevel duplication
Replace the copy-pasted getZoomLevel() + manual zoom-division math in
GameLayout, ReplayOverlay, and InGame with a shared computeDropCoords()
helper. The helper derives the CSS↔visual-pixel scale factor from the
container element itself (offsetWidth / getBoundingClientRect().width)
rather than reading document.documentElement.zoom, which only captures
CSS zoom — not browser pinch-zoom or transform:scale.
