+++ src/themes/redesign/screens/InGame.jsx
@@ -1,14 +1,11 @@
 import { useRef, useState, useCallback } from 'react';
 import { useGame } from '../../../context/GameContext.jsx';
 import { GRID_COLS, GRID_ROWS } from '../../../utils/gameConstants.js';
+import { computeDropCoords } from '../../../utils/dropCoordUtils.js';
 import { A } from '../../../utils/actionTypes.js';
 import Shell from '../components/Shell.jsx';
 import Board from '../components/Board.jsx';
 import DroppingOverlay from '../../../shared/overlays/DroppingOverlay.jsx';
-
-function getZoomLevel() {
-  return parseFloat(window.getComputedStyle(document.documentElement).zoom) || 1;
-}
 
 function InGame({ onHowToPlay, onLogin, onSettings }) {
   const { state, dispatch } = useGame();
@@ -48,15 +45,7 @@ function InGame({ onHowToPlay, onLogin, onSettings }) {
       return;
     }
 
-    const zoom = getZoomLevel();
-    const containerRect = screenEl.getBoundingClientRect();
-    const fromRect = fromEl.getBoundingClientRect();
-    const gridRect = gridEl.getBoundingClientRect();
-    const colWidth = gridRect.width / GRID_COLS;
-    const rowHeight = gridRect.height / GRID_ROWS;
-    const cellSize = Math.min(colWidth, rowHeight) / zoom;
-    const colLeft = gridRect.left + column * colWidth + (colWidth - cellSize * zoom) / 2;
-
+    const coords = computeDropCoords(screenEl, fromEl, gridEl, column, destRow);
     const id = `${Date.now()}-${Math.random()}`;
     inFlightColumnsRef.current.set(column, cif + 1);
     inFlightCountRef.current++;
@@ -65,10 +54,7 @@ function InGame({ onHowToPlay, onLogin, onSettings }) {
       id,
       column,
       letter: letter.char,
-      from:    { x: (fromRect.left - containerRect.left) / zoom, y: (fromRect.top - containerRect.top) / zoom },
-      toTop:   { x: (colLeft - containerRect.left) / zoom, y: (gridRect.top - containerRect.top) / zoom },
-      toFinal: { x: (colLeft - containerRect.left) / zoom, y: (gridRect.top + destRow * rowHeight - containerRect.top) / zoom },
-      cellSize,
+      ...coords,
     }]);
   }, [state.nextQueue, getDestRow, dispatch]);
